feature/CP-929 (#9)
Definition of Done

PR merged into develop

The Policy we create with the RoleArn has permissions just for the Master Payer Account Type. Please use the role-access-policy-template-master-org-account.json file in the CLI as reference, though this file has too many permissions. The Policy we create in here should have all the permissions from that file around Billing, e.g. cur:, ce:, etc. and permissions to read data from all Services that have reserved capacity (you may have to search to find these services).

Interface

Input

IsMasterPayerAccount: 

Create a condition on when this is `“true”` and MasterPayerBillingBucketName is not `“null”`

All resources in the template are Conditional on it

MasterPayerBillingBucketName:

Create a condition on when this is not `“null”` and IsMasterPayerAccount is "true”

All resources in the template are Conditional on it

Use this for SNSTopicPolicy and Queue Subscription

ReactorAccountId: Use this for cross account Role Principal

ExternalId: Use this for cross account Role Trust Permissions

Output

RoleArn: The cross account role created

Resource Tags

All Taggable Resources should have these tags

  Tags:
    - Key: cloudzero-stack
      Value: !Ref AWS::StackName
    - Key: cloudzero-reactor-account-id
      Value: !Ref ReactorAccountId